requirements.txt
python-telegram-bot[job-queue]==21.6
openai>=1.40.0
httpx>=0.27.0
flask>=3.0.0
